Teste de arquivos pequenos, grandes e realmente grandes funcionando assim como tratamento de várias requisições

+++ utils/connection_item.h
@@ -15,14 +15,14 @@ enum ConnectionStates
 
 typedef struct ConnectionStruct
 {
-  int32_t socket_descriptor;
+  int32_t  socket_descriptor;
   uint8_t  state;
   uint8_t  header_sent;
-  int32_t wroteData;
-  int32_t response_size;
-  char    *request;
-  FILE    *resource_file;
-  char    *header;
+  uint64_t wroteData;
+  uint64_t response_size;
+  char     *request;
+  FILE     *resource_file;
+  char     *header;
 
   /*List*/
   struct ConnectionStruct *previous_ptr;
